feat: useCommitteeHeader 훅 구현
